@@ -1,42 +1,42 @@
-FactoryGirl.define do
-  factory :user do
-    name 'Matt Hearnes'
-    sequence(:email) do |n| "matt#{n}.matt@hearnes.com" end
-    password 'OurLadyOfKnock'
-    password_confirmation 'OurLadyOfKnock'
+FactoryBot.define do
+  # factory :user do
+  #   name 'Matt Hearnes'
+  #   sequence(:email) do |n| "matt#{n}.matt@hearnes.com" end
+  #   password 'OurLadyOfKnock'
+  #   password_confirmation 'OurLadyOfKnock'
 
-    trait :activated do
-      activated true
-      activated_at { Time.now }
-    end
+  #   trait :activated do
+  #     activated true
+  #     activated_at { Time.now }
+  #   end
 
-    trait :approved do
-      approved true
-      approved_at { Time.now }
-    end
+  #   trait :approved do
+  #     approved true
+  #     approved_at { Time.now }
+  #   end
 
-    trait :active do
-      active true
-    end
+  #   trait :active do
+  #     active true
+  #   end
 
-    trait :inactive do
-      active false
-    end
+  #   trait :inactive do
+  #     active false
+  #   end
 
-    factory :admin do
-      name 'Mr. Admin'
-      sequence(:email) do |n| "admin#{n}@example.com" end
-      activated
-      approved
-      admin true
-    end
+  #   factory :admin do
+  #     name 'Mr. Admin'
+  #     sequence(:email) do |n| "admin#{n}@example.com" end
+  #     activated
+  #     approved
+  #     admin true
+  #   end
 
-    factory :remembered_user do
-      remember_digest 'RememberMe'
-    end
+  #   factory :remembered_user do
+  #     remember_digest 'RememberMe'
+  #   end
 
-    factory :inactive_user do
-      inactive
-    end
-  end
+  #   factory :inactive_user do
+  #     inactive
+  #   end
+  # end
 end
