@@ -2,7 +2,7 @@ class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
   include SessionsHelper
 
-  before_action :prepare_meta_tags, if: "request.get?"
+  before_action :prepare_meta_tags, if: -> { request.get? }
 
   def prepare_meta_tags(options={})
     site_name   = "Saint Stories"
