@@ -11,12 +11,9 @@ module SaintStories
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 5.1
 
-    # Settings in config/environments/* take precedence
-    # over those specified here.
-    # Application configuration should go into files
-    # in config/initializers
-    # -- all .rb files in that directory are automatically loaded.
-    # Include the authenticity token in remote forms.
-    config.action_view.embed_authenticity_token_in_remote_forms = true
+    # Settings in config/environments/* take precedence over those specified here.
+    # Application configuration can go into files in config/initializers
+    # -- all .rb files in that directory are automatically loaded after loading
+    # the framework and any gems in your application.
   end
 end
